refactor: extract getStepDefaultDefinition function to its own file
